Move reporting & info to separate packages
Addresses circular dependency issues detected during type checks.
