+++ packages/kbn-scout-info/README.md
@@ -1,0 +1,6 @@
+# @kbn/scout-info
+
+This package contains stores information that's commonly used by packages in the `@kbn/scout*` namespace, and any other
+ package that wishes to extend Scout functionality.
+
+Check out the `@kbn/scout` package if you want to learn more about Scout.
