+++ packages/kbn-scout-info/kibana.jsonc
@@ -1,0 +1,6 @@
+{
+  "type": "shared-common",
+  "id": "@kbn/scout-info",
+  "owner": "@elastic/appex-qa",
+  "devOnly": true
+}
