+++ packages/kbn-scout-info/src/paths.ts
