+++ packages/kbn-scout-reporting/README.md
@@ -1,0 +1,5 @@
+# @kbn/scout-reporting
+
+This package contains reporting functionality for Scout.
+
+Check out the `@kbn/scout` package if you want to learn more about Scout.
