+++ packages/kbn-scout-reporting/kibana.jsonc
@@ -1,0 +1,6 @@
+{
+  "type": "shared-common",
+  "id": "@kbn/scout-reporting",
+  "owner": "@elastic/appex-qa",
+  "devOnly": true
+}
