+++ packages/kbn-scout-reporting/src/cli/common.ts
