+++ packages/kbn-scout-reporting/src/cli/initialize_report_datastream.ts
