+++ packages/kbn-scout-reporting/src/cli/upload_events.ts
