+++ packages/kbn-scout-reporting/src/reporting/playwright.ts
@@ -20,7 +20,7 @@ import type {
 
 import path from 'node:path';
 import { ToolingLog } from '@kbn/tooling-log';
-import { SCOUT_REPORT_OUTPUT_ROOT } from '../paths';
+import { SCOUT_REPORT_OUTPUT_ROOT } from '@kbn/scout-info';
 import { generateTestRunId, getTestIDForTitle, ScoutReport, ScoutReportEventAction } from '.';
 import { environmentMetadata } from '../datasources';
