+++ packages/kbn-scout-reporting/src/reporting/report/event.ts
