+++ packages/kbn-scout-reporting/src/reporting/report/persistence/component_templates.ts
