+++ packages/kbn-scout-reporting/src/reporting/report/persistence/index_templates.ts
