+++ packages/kbn-scout-reporting/src/reporting/report/persistence/mappings.ts
