+++ packages/kbn-scout-reporting/tsconfig.json
@@ -17,17 +17,6 @@
     "@kbn/tooling-log",
     "@kbn/dev-cli-runner",
     "@kbn/dev-cli-errors",
-    "@kbn/ci-stats-reporter",
-    "@kbn/repo-info",
-    "@kbn/es",
-    "@kbn/dev-proc-runner",
-    "@kbn/test",
-    "@kbn/es-archiver",
-    "@kbn/dev-utils",
-    "@kbn/mock-idp-utils",
-    "@kbn/test-suites-xpack",
-    "@kbn/test-subj-selector",
-    "@kbn/scout-info",
-    "@kbn/scout-reporting"
+    "@kbn/scout-info"
   ]
 }
