+++ packages/kbn-test/tsconfig.json
@@ -37,6 +37,6 @@
     "@kbn/core-saved-objects-api-server",
     "@kbn/mock-idp-utils",
     "@kbn/code-owners",
-    "@kbn/scout",
+    "@kbn/scout-reporting",
   ]
 }
