+++ tsconfig.base.json
@@ -1546,6 +1546,10 @@
       "@kbn/saved-search-plugin/*": ["src/plugins/saved_search/*"],
       "@kbn/scout": ["packages/kbn-scout"],
       "@kbn/scout/*": ["packages/kbn-scout/*"],
+      "@kbn/scout-info": ["packages/kbn-scout-info"],
+      "@kbn/scout-info/*": ["packages/kbn-scout-info/*"],
+      "@kbn/scout-reporting": ["packages/kbn-scout-reporting"],
+      "@kbn/scout-reporting/*": ["packages/kbn-scout-reporting/*"],
       "@kbn/screenshot-mode-example-plugin": ["examples/screenshot_mode_example"],
       "@kbn/screenshot-mode-example-plugin/*": ["examples/screenshot_mode_example/*"],
       "@kbn/screenshot-mode-plugin": ["src/plugins/screenshot_mode"],
